Small refactor.
The LoadChunks coroutine has been moved out of the manager and into the individual chunk loading objects in preperation for multiple objects to be able to chunkload at once.
