+++ Assets/Scripts/ChunkManager.cs
@@ -3,8 +3,6 @@
 // PURPOSE: Manages chunk data, such as which chunks are currently loaded
 
 using UnityEngine;
-using System.Collections;
-using System;
 
 public class ChunkManager : MonoBehaviour
 {
@@ -18,11 +16,6 @@ public class ChunkManager : MonoBehaviour
     int renderDistance = 4;
 
     public int numberOfLoaders = 0;
-    int currentLoaderID = 0;
-
-    //Coroutine
-    IEnumerator currentLoadCoroutine;
-    bool bIsGenerating = true;
 
     void Start()
     {
@@ -40,11 +33,10 @@ public class ChunkManager : MonoBehaviour
         return numberOfLoaders;
     }
 
-    public void GetChunkBelowObject(GameObject m_ObjectToRaycast, int m_RenderDistance, int m_LoaderID)
+    public GameObject GetChunkBelowObject(GameObject m_ObjectToRaycast, int m_RenderDistance, int m_LoaderID)
     {
         //Set render distance
         renderDistance = m_RenderDistance;
-        currentLoaderID = m_LoaderID;
 
         //Check what is below the object calling the function
         RaycastHit hit;
@@ -56,35 +48,26 @@ public class ChunkManager : MonoBehaviour
         //Optimisation
         if(hitObject == previousObject)
         {
-            return;
+            return null;
         }
 
         //Set previous object to hitobject
         previousObject = hitObject;
-
-        GetSurroundingChunks();
-
-        //Run loading coroutine
-        if(currentLoadCoroutine != null)
-        {
-            StopCoroutine(currentLoadCoroutine);
-        }
-        currentLoadCoroutine = LoadChunks();
-        StartCoroutine(currentLoadCoroutine);
+        return previousObject;
+        //chunknames = GetSurroundingChunks(new Vector2(previousObject.transform.position.x, previousObject.transform.position.z));
     }
 
-    void GetSurroundingChunks()
+    public string[] GetSurroundingChunks(Vector2 chunkPos)
     {
         //Gets current chunk coordinate
-        int xPos = (int) previousObject.transform.position.x / xChunkSize;
-        int zPos = (int) previousObject.transform.position.z / zChunkSize;
+        int xPos = (int) chunkPos.x / xChunkSize;
+        int zPos = (int) chunkPos.y / zChunkSize;
 
         //Names of chunks to be loaded
-        chunknames = new string[renderDistance * renderDistance * numberOfLoaders];
+        string[] chunknames = new string[renderDistance * renderDistance * numberOfLoaders];
 
         //Loop for filling chunknames array based on surrounding chunks
-        int k = 0;
-        for (int z = 0; z < renderDistance; z++)
+        for (int z = 0, k = 0; z < renderDistance; z++)
         {
             for (int x = 0; x < renderDistance; x++)
             {
@@ -92,59 +75,7 @@ public class ChunkManager : MonoBehaviour
                 k++;
             }
         }
-    }
-    
-    IEnumerator LoadChunks()
-    {
-        //For each chunk
-        for (int i = 0; i < mg.transform.childCount; i++)
-        {
-            //For each element in the chunk array
-            for (int j = 0; j < chunknames.Length; j++)
-            {
-                //If chunk does not exist, create a new one at this position. This allows creation of endless terrain.
-                GameObject chunk = GameObject.Find(chunknames[j]);
-                if (chunk == null)
-                {
-                    string chunkname = chunknames[j];
 
-                    //Get coordinates of chunk from name of chunk
-                    int a = chunkname.IndexOf("(");
-                    int b = chunkname.IndexOf(",");
-                    int x = int.Parse(chunkname.Substring(a + 1, (b - a) - 1));
-
-                    int c = chunkname.IndexOf(")");
-                    int z = int.Parse(chunkname.Substring(b + 2, (c - (b + 1)) - 1));
-
-                    //Create new chunk at these coordinates
-                    mg.CreateNewChunk(x, z, currentLoaderID);
-
-                    //Coroutine pause point
-                    if (!bIsGenerating)
-                    {
-                        yield return null;
-                    }
-                }
-            }
-            //Removes chunks that are not loaded
-            RemoveUnloadedChunks(i);
-        }
-        bIsGenerating = false;
-    }
-
-    //Function to remove unloaded chunks
-    void RemoveUnloadedChunks(int index)
-    {
-        //Get position of current transform in array
-        string value = mg.transform.GetChild(index).name;
-        int pos = Array.IndexOf(chunknames, value);
-
-        //If transform is not in chunknames for loaded chunks
-        if (pos == -1)
-        {
-            //Destroy the current object
-            Destroy(mg.transform.GetChild(index).gameObject);
-
-        }
+        return chunknames;
     }
 }
